pyproject.toml
[tool.poetry]
name = "ai-security-engineer-take-home"
version = "0.1.0"
description = ""
authors = ["4L3xD <alessandra.olea@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.4.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
